refactor: extract BombaGameCore shared module, eliminate logic duplication
Move all pure game logic from server/src/games/bomba.js and client/src/local/games/LocalBombaGame.js into a new shared/BombaGameCore.js base class. Both BombaGame (server) and LocalBombaGame (client) now extend BombaGameCore and only override emitState(). Add @shared Vite alias for client imports.
